backend/pyproject.toml
[tool.poetry]
name = "diary-workout-tracker-backend"
version = "0.1.0"
description = ""
authors = [
    "Струнникова Елизавета <liza.strunnikova@yandex.ru>",
    "Оганин Петр <necroshizo@gmail.com>",
    "Шапиро Алексей <novosib@gmail.com>",
    "Шлемин Сергей <schlyomin@yandex.ru",
    "Яременко Владислав <vlad.yaremenko.98@yandex.ru>",
    "Ястребов Владислав <blakkheart66@gmail.com>",
]
readme = "../README.md"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".mypy_cache",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "venv",
    "frontendruff",
    "migrations",
]
target-version = "py39"
line-length = 120
lint.ignore = ["D"]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
docstring-code-format = true


[tool.pytest.ini_options]
pythonpath = "backend"
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "tests.py test_*.py *_tests.py"


[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.1"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.1"
ruff = "^0.2.1"
pytest = "^8.0.0"
pytest-django = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
